requirements.txt
Flask
openai
python-dotenv
gunicorn

